Fix the bug with the report date
Technically not a bug, if there is not a report that includes "today"
the report will only show the "first" report's range. I changed it to be
the "last" report found, which is more likely to be accurate. Long term
I need to make that selectable, from the front screen.